% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisotropic_diffusion.R
\name{anisotropic_diffusion}
\alias{anisotropic_diffusion}
\title{Anistropic diffusion}
\usage{
anisotropic_diffusion(x, iteration = 50, lambda = 0.2, k = 10)
}
\arguments{
\item{x}{a grayscale matrix with values ranging in [0,1] or [0, 255]. Rasters from package raster
and terra are also supported}

\item{iteration, lambda}{numeric}

\item{k}{integer}
}
\description{
In image processing and computer vision, anisotropic diffusion, is a technique aiming at reducing
image noise without removing significant parts of the image content, typically edges, lines or
other details that are important for the interpretation of the image.
}
\examples{
f = system.file("extdata", "albert.png", package="anidiff")
m1 = t(apply(png::readPNG(f)[,,1], 2, rev))
m2 = anisotropic_diffusion(m1)
image(m1, col = gray((0:255)/255))
image(m2, col = gray((0:255)/255))
}
